requirements.txt
#requests==2.31.0
future==1.0.0
obspy==1.4.1
rdflib==7.0.0
alive_progress==3.1.4
python-dotenv==1.0.1